requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
google-genai>=0.1.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.0
asyncpg>=0.29.0
passlib[bcrypt]>=1.7.4
itsdangerous>=2.1.2
httpx>=0.27.0
python-multipart>=0.0.9
fastapi-sso>=0.15.0
python-pptx>=0.6.21
fpdf2>=2.7.1
python-docx>=1.1.0
